requirements.txt
Pillow>=10.0.0
ultralytics>=8.0.0
torch>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
customtkinter>=5.0.0
transformers>=4.30.0
llama-cpp-python>=0.2.0
onnxruntime>=1.15.0
huggingface_hub>=0.20.0
pandas>=2.0.0

